EvalRunner/pyproject.toml
[tool.poetry]
name = "reactbench"
version = "0.1.0"
description = "An eval to test react coding capabilities for modern LLMS"
authors = ["Kshitij Banerjee <ksh.dce@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.30.1"
anthropic = "^0.25.9"
load-dotenv = "^0.1.0"
tenacity = "^8.3.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
